requirements.txt
cycler==0.11.0
et-xmlfile==1.1.0
gevent==22.10.2
greenlet==2.0.2
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.24.2
openpyxl==3.1.1
pandas==1.1.5
Pillow==8.4.0
pyodbc==4.0.35
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7.1
pyzmq==24.0.1
rel==0.4.7
six==1.16.0
sysv-ipc==1.1.0
tzdata==2023.3
websocket==0.2.1
websocket-client==1.3.1
websockets==9.1
xlwings==0.27.15
zmq==0.0.0
zope.event==4.6
zope.interface==6.0
